requirements.txt
accelerate==1.8.1
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.9
charset-normalizer==2.0.12
click==8.2.1
colorama==0.4.6
diffusers==0.27.2
filelock==3.13.1
Flask==2.0.1
fsspec==2024.6.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.23.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==25.0
pillow==11.0.0
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==0.19.0
PyYAML==6.0.2
regex==2024.11.6
replicate==0.20.0
requests==2.26.0
safetensors==0.5.3
sniffio==1.3.1
sympy==1.13.3
tokenizers==0.19.1
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.41.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==1.26.20
Werkzeug==2.0.3
zipp==3.23.0
